+++ enlighten/file_io/LoggingFeature.py
@@ -42,8 +42,9 @@ class LoggingFeature:
 
         # if verbose logging was specified at the command-line OR 
         # previously set via .ini, use that
-        if log.isEnabledFor(logging.DEBUG) or self.ctl.config.get_bool("logging", "verbose"):
+        if self.ctl.log_level == "DEBUG":
             self.cb_verbose.setChecked(True)
+            log.debug("checked verbose logging because specified on command-line")
 
         try:
             offset_file = applog.get_location() + ".offset"
@@ -63,12 +64,10 @@ class LoggingFeature:
         if enabled:
             log.info("enabling verbose logging")
             logging.getLogger().setLevel(logging.DEBUG)
-            self.ctl.config.set("logging", "verbose", "True")
             self.ctl.log_level = "DEBUG"
         else:
             log.info("disabling verbose logging")
             logging.getLogger().setLevel(logging.INFO)
-            self.ctl.config.set("logging", "verbose", "False")
             self.ctl.log_level = "INFO"
 
     def paused(self):